found a workaround to ctype limitations, begun json_parse_number spec

+++ src/ion/functions/json_parse_number.h
@@ -1,3 +1,3 @@
-static inline u64 json_parse_number (
+static inline int json_parse_number (
     struct io* source
 );
